Unify typedb and surrealdb test
