+++ tests/surrealdb/unit/mutations/basic.test.ts
@@ -1,0 +1,4 @@
+import { testBasicMutation } from "../../../unit/mutations/basic";
+import { init } from "../../helpers/init";
+
+testBasicMutation(init);
