+++ tests/surrealdb/unit/mutations/batched.test.ts
@@ -1,0 +1,4 @@
+import { testBatchedMutation } from "../../../unit/mutations/batched";
+import { init } from "../../helpers/init";
+
+testBatchedMutation(init);
