+++ tests/surrealdb/unit/mutations/errors.test.ts
@@ -1,0 +1,4 @@
+import { testMutationErrors } from "../../../unit/mutations/errors";
+import { init } from "../../helpers/init";
+
+testMutationErrors(init);
