+++ tests/typedb/helpers/init.ts
@@ -1,0 +1,14 @@
+import { schema } from '../../mocks/schema';
+import { config } from '../mocks/config';
+import { setup } from '../../helpers/setup';
+
+export const init = async () => setup({
+		config,
+		schema,
+		tqlPathMap: {
+			default: {
+				schema: './tests/typedb/mocks/schema.tql',
+				data: './tests/typedb/mocks/data.tql',
+			},
+		}
+	});
