+++ tests/typedb/unit/mutations/edges.test.ts
@@ -1,0 +1,4 @@
+import { testEdgesMutation } from "../../../unit/mutations/edges";
+import { init } from "../../helpers/init";
+
+testEdgesMutation(init);
