+++ tests/typedb/unit/mutations/prehook.test.ts
@@ -1,0 +1,4 @@
+import { testMutationPrehooks } from "../../../unit/mutations/preHooks";
+import { init } from "../../helpers/init";
+
+testMutationPrehooks(init);
