+++ tests/typedb/unit/mutations/replaces.test.ts
@@ -1,0 +1,4 @@
+import { testReplaceMutation } from "../../../unit/mutations/replaces";
+import { init } from "../../helpers/init";
+
+testReplaceMutation(init);
