+++ tests/typedb/unit/mutations/unsupported.test.ts
@@ -1,0 +1,4 @@
+import { testUnsupportedMutation } from "../../../unit/mutations/unsupported";
+import { init } from "../../helpers/init";
+
+testUnsupportedMutation(init);
